requirements.txt
pandas>=1.4.0,<=2.2.3
xarray>=0.21.0,<=2024.11.0
numpy>=1.19.3,<=1.26.4
matplotlib>=3.4.0,<=3.6.1
Pillow>=9.0.0,<=10.2.0
scikit-learn>=1.0.0,<=1.6.0
seaborn>=0.11.1
openpyxl>=3.0.7,<=3.0.9
Jinja2>=3.0.0,<=3.1.4
